Item11267: Add handler to set Security ACLs
This adds an beforeSaveHandler that examines the form data.   If the
topic is set to the "Security" priority and is not in a closed state (No
Action or Closed), then the ALLOWTOPICVIEW ACL is set to restrict access
to the SecurityGroup, AdminGroup, the reporter and anyone identified in
WaitingFor.

When the task is closed, the prior version is examined to see if it was
a Security task.  If it was, then the ACL's are removed.   If not, the
ACLs are left as-is.

git-svn-id: http://svn.foswiki.org/trunk/FoswikiOrgPlugin@17752 0b4bb1d4-4e5a-0410-9cc4-b2b747904278