Item11267: Add support for Item specific branches.
Checkins to any branch named ItemNNNNNN (2-6 digits)  with any suffix will now be
processed when the web push occurs.

 - Full branch name is recorded in CheckinsOnBranches
 - Git hash refs recorded into ItemBranchCheckins

Note that the task update is not restricted to the task prefixed in
the branch-name.  All tasks in the commit message will be updated with
the branch specific checkin information.

I'm not sure what will happen once the branch is merged into master.
The commit hashes might become dead links.